work

+++ src/modules/card-module/component-card-list/component-card-list.component.ts
@@ -1,50 +1,68 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Router } from '@angular/router';
-import { FormBuilder, FormControl, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
-
-const f = FormGroup;
 
 @Component({
   selector: 'app-component-card-list',
   templateUrl: './component-card-list.component.html',
-  styleUrls: ['./component-card-list.component.scss']
+  styleUrls: [],
 })
-
 export class ComponentCardListComponent implements OnInit {
   data: any[] = [];
   newCard: any = {};
   cards: any[] = [];
   selectedCard: any;
+  token: string | null = null;
 
   constructor(private http: HttpClient, private router: Router) {}
 
-ngOnInit() {
-  this.http.get<any[]>('http://localhost:3000/api/cards')
-  .subscribe(response => {
-  this.data = response;
-  console.log(this.data);
-  });
-}
+  ngOnInit() {
+    this.token = localStorage.getItem('token');
 
-onSubmit() {
-  this.http.put('http://localhost:3000/api/cards', this.newCard)
-  .subscribe(response => {
-  this.data.push(response);
-  this.newCard = {};
-  window.location.reload();
-  });
-}
+    if (this.token) {
+      this.http
+        .get<any[]>('http://localhost:3000/api/cards')
+        .subscribe((response) => {
+          this.data = response;
+          console.log(this.data);
+        });
+    } else {
+      this.router.navigateByUrl('/login');
+    }
+  }
 
-deleteCard(id: string) {
-  const url = `http://localhost:3000/api/cards/${id}`;
-  return this.http.delete(url).subscribe(() => {
-  // this.cards = this.cards.filter(card => card.id !== id);
-  window.location.reload();
-  });
-}
+  onSubmit() {
+    if (this.token) {
+      this.http
+        .put('http://localhost:3000/api/cards', this.newCard)
+        .subscribe((response) => {
+          this.data.push(response);
+          this.newCard = {};
+          window.location.reload();
+        });
+    } else {
+      this.router.navigateByUrl('/login');
+    }
+  }
+
+  deleteCard(id: string) {
+    if (this.token) {
+      const url = `http://localhost:3000/api/cards/${id}`;
+      return this.http.delete(url).subscribe(() => {
+        // this.cards = this.cards.filter(card => card.id !== id);
+        window.location.reload();
+      });
+    } else {
+      this.router.navigateByUrl('/login');
+      return undefined; // Ajouter une valeur de retour par défaut
+    }
+  }
 
   showDetails(id: string) {
-    this.router.navigateByUrl(`/card/${id}`);
+    if (this.token) {
+      this.router.navigateByUrl(`/card/${id}`);
+    } else {
+      this.router.navigateByUrl('/login');
+    }
   }
 }
